fix: rename footer link
