+++ packages/desktop/src/index.ts
@@ -10,6 +10,7 @@ export * from './link'
 export * from './switch'
 export * from './breadcrumbs'
 export * from './nav-link'
+export * from './breadcrumbs'
 
 export {TextField as MaskTextField} from './text-field'
 export {PasswordField as MaskPasswordField} from './password-field'
