+++ packages/mobile/src/index.ts
@@ -9,6 +9,7 @@ export * from './radio-field'
 export * from './typography'
 export * from './link'
 export * from './nav-link'
+export * from './breadcrumbs'
 
 export {TextField as MaskTextField} from './text-field'
 export {PasswordField as MaskPasswordField} from './password-field'
